ui fix RequestViewer
